+++ notification/src/main/java/com/team15gijo/notification/application/dto/v1/message/CommentNotificationEventDto.java
@@ -1,4 +1,4 @@
-package com.team15gijo.notification.application.dto.v1;
+package com.team15gijo.notification.application.dto.v1.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CommentNotificationEvent {
+public class CommentNotificationEventDto {
     private Long receiverId;         // 알림 받을 사용자 ID
     private String nickname;           // 알림을 보내는 사용자의 닉네임
     private String commentContent;
