+++ notification/src/main/java/com/team15gijo/notification/application/message/KafkaTestProducer.java
@@ -1,19 +1,19 @@
 package com.team15gijo.notification.application.message;
 
-import com.team15gijo.notification.application.dto.v1.CommentNotificationEvent;
+import com.team15gijo.notification.application.dto.v1.message.CommentNotificationEventDto;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
 @Service
 public class KafkaTestProducer {
 
-    private final KafkaTemplate<String, CommentNotificationEvent> kafkaTemplate;
+    private final KafkaTemplate<String, CommentNotificationEventDto> kafkaTemplate;
 
     public KafkaTestProducer(KafkaTemplate kafkaTemplate) {
         this.kafkaTemplate = kafkaTemplate;
     }
 
-    public void sendCommentCreate(CommentNotificationEvent event) {
+    public void sendCommentCreate(CommentNotificationEventDto event) {
         kafkaTemplate.send("COMMENT", event);
     }
